refactor(models): partition session command events by domain (#4136)
Slice B splits session_command_events into domain libraries under
app_events/session_* with a thin barrel; clears the physical-size
grandfather entry and adds barrel smoke tests.

Refs #4136
